Minor fixes here and there

+++ zsh/aliases.zsh
@@ -32,8 +32,8 @@ nextd() {
 }
 
 prevd() {
-  prev_dirs+=("$PWD")
-  popd "$@"
+  popd "$@" && 
+    prev_dirs+=("$OLDPWD")
 }
 
 zle_nextd() {
@@ -136,12 +136,12 @@ ghd() {
     return
   fi
 
-  mkdir -p "$dst" \
-  && if git clone --recursive "https://github.com/$repo_path" "$dst"; then
-    cd "$dst"
-  else
-    rm -Rf "$dst"
-  fi
+  mkdir -p "$dst" &&
+    if git clone --recursive "https://github.com/$repo_path" "$dst"; then
+      cd "$dst"
+    else
+      rm -Rf "$dst"
+    fi
 }
 
 gh() {
